requirements.txt
pdf2image==1.17.0
PyPDF2==3.0.1
python-dotenv==1.1.0
requests==2.32.3
